moved components slighltly
